Refactor `runtime` code structure
