+++ runtime/common/src/constants.rs
@@ -1,49 +1,5 @@
-// This file is part of Substrate.
-
-// Copyright (C) Parity Technologies (UK) Ltd.
-// SPDX-License-Identifier: Apache-2.0
-
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-// 	http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-//! A set of constant values used in substrate runtime.
-
-/// Money matters.
-pub mod currency {
-	pub type Balance = u128;
-
-	/// Constant values used within the runtime.
-	pub const MICROPAN: Balance = 1_000_000;
-	pub const MILLIPAN: Balance = 1_000 * MICROPAN;
-	pub const CENTS: Balance = 10 * MILLIPAN; // assume this is worth about a cent.
-	pub const PAN: Balance = 100 * CENTS;
-
-	pub const INIT_SUPPLY_FACTOR: Balance = 100;
-	pub const STORAGE_BYTE_FEE: Balance = 4 * MICROPAN * INIT_SUPPLY_FACTOR;
-
-	/// Charge fee for stored bytes and items.
-	pub const fn deposit(items: u32, bytes: u32) -> Balance {
-		items as Balance * 15 * CENTS + (bytes as Balance) * STORAGE_BYTE_FEE
-	}
-
-	/// Charge fee for stored bytes and items as part of `pallet-contracts`.
-	pub const fn contracts_deposit(items: u32, bytes: u32) -> Balance {
-		items as Balance * MILLIPAN * INIT_SUPPLY_FACTOR + (bytes as Balance) * STORAGE_BYTE_FEE
-	}
-}
-
 /// Time.
 pub mod time {
-
 	/// Since BABE is probabilistic this is the average expected block time that
 	/// we are targeting. Blocks will be produced at a minimum duration defined
 	/// by `SLOT_DURATION`, but some slots will not be allocated to any
@@ -62,8 +18,7 @@ pub mod time {
 	///
 	/// <https://research.web3.foundation/en/latest/polkadot/block-production/Babe.html#-6.-practical-results>
 
-	pub type BlockNumber = u32;
-	pub type Moment = u64;
+	use crate::{BlockNumber, Moment};
 
 	pub const MILLISECS_PER_BLOCK: Moment = 3000;
 	pub const SECS_PER_BLOCK: Moment = MILLISECS_PER_BLOCK / 1000;
