+++ runtime/local/src/pallets/election_provider_multi_phase.rs
@@ -2,10 +2,11 @@
 
 use codec::Decode;
 use frame_support::{
+	dispatch::DispatchClass,
 	traits::ConstU32,
 	weights::constants::BlockExecutionWeight,
 };
-use sp_runtime::transaction_validity::TransactionPriority;
+use sp_runtime::{transaction_validity::TransactionPriority, Perbill};
 
 use crate::*;
